moved around a ton of files and deleted a bunch, if something breaks now, go to commit 1 or 2 before this to see if it works
